# swagger.yaml
openapi: 3.0.0
info:
  title: API d'Authentification et Gestion des Utilisateurs
  version: 1.0.0
  description: API pour l'authentification et la gestion des utilisateurs.
servers:
  - url: http://localhost:4000/api/v1/
    description: Serveur de développement local
...
paths:
...
  /works:
    post:
      tags:
        - Travaux (Admin seulement)
      summary: Ajouter un nouveau travail
      description: Ajoute un travail avec un titre, une image de couverture et des fichiers (images/vidéos).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Titre du travail
                cover:
                  type: string
                  format: binary
                  description: Image de couverture
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Images ou vidéos associées
                date:
                  type: string
                  format: date-time
                  description: Date de l'événement ou du travail
      responses:
        "201":
          description: Travail ajouté avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Work"
        "400":
          description: Données invalides
    get:
      tags:
        - Travaux (Admin seulement)
      summary: Récupérer la liste des travaux
      description: Renvoie la liste de tous les travaux (titre et cover pour le slider).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liste des travaux
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Work"
  /works/{id}:
    get:
      tags:
        - Travaux (Admin seulement)
      summary: Récupérer le détail d'un travail
      description: Renvoie toutes les informations d'un travail (titre, cover, fichiers…).
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID du travail à récupérer
      responses:
        "200":
          description: Détail du travail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Work"
        "404":
          description: Travail non trouvé
    delete:
      tags:
        - Travaux (Admin seulement)
      summary: Supprimer un travail (administrateur uniquement)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID du travail à supprimer
      responses:
        "200":
          description: Travail supprimé avec succès
        "401":
          description: Non autorisé
        "403":
          description: Accès refusé
        "404":
          description: Travail non trouvé
    put:
      tags:
        - Travaux (Admin seulement)
      summary: Modifier un travail
      description: Modifie le titre, la cover ou les fichiers associés à un travail (admin uniquement).
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID du travail à modifier
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Nouveau titre du travail
                cover:
                  type: string
                  format: binary
                  description: Nouvelle image de couverture (optionnel)
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Nouvelles images ou vidéos associées (optionnel)
      responses:
        "200":
          description: Travail modifié avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Work"
        "400":
          description: Données invalides
        "404":
          description: Travail non trouvé
        "500":
          description: Erreur serveur
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Work:
      type: object
      properties:
        id:
          type: string
          description: Identifiant unique du travail (ObjectId)
          example: "665f1f77bcf86cd799439011"
        title:
          type: string
          description: Titre du travail
          example: "Réflexologie plantaire"
        cover:
          type: string
          description: URL de l'image principale (cover)
          example: "/uploads/cover1.jpg"
        files:
          type: array
          description: Liste des fichiers (images/vidéos) associés
          items:
            type: object
            properties:
              type:
                type: string
                enum: [image, video]
                description: Type du fichier
                example: "image"
              url:
                type: string
                description: URL du fichier
                example: "/uploads/photo1.jpg"
        date:
          type: string
          format: date-time
          description: Date de l'événement ou du travail
          example: "2024-07-01T18:00:00.000Z"